+++ cypress/e2e/Colonnes/TableauStudies.cy.ts
@@ -52,7 +52,7 @@ describe('Page des études - Colonnes du tableau', () => {
     cy.get('thead[class="ant-table-thead"]')
       .find('th[class*="ant-table-cell"]').eq(8)
       .should('have.class', 'ant-table-column-has-sorters')
-      .contains('Biospecimen').should('exist');
+      .contains('Biospecimens').should('exist');
 
     cy.get('thead[class="ant-table-thead"]')
       .find('th[class*="ant-table-cell"]').eq(9)