+++ cypress/e2e/Telechargement/TableauxPageParticipant_4.cy.ts
@@ -7,6 +7,7 @@ beforeEach(() => {
   cy.removeFilesFromFolder(Cypress.config('downloadsFolder'));
   cy.login();
   cy.visitParticipantEntity('pt-0dxdyebh');
+  cy.get('[id="biospecimen"] [data-node-key="table"]').clickAndWait({force: true});
   cy.resetColumns('biospecimen');
   cy.get('div[id="content"] svg[data-icon="download"]').eq(5).clickAndWait({force:true});
   cy.waitUntilFile(oneMinute);