+++ src/App.tsx
@@ -8,6 +8,7 @@ import loadable from '@loadable/component';
 import { useKeycloak } from '@react-keycloak/web';
 import { ConfigProvider } from 'antd';
 import enUS from 'antd/lib/locale/en_US';
+import esES from 'antd/lib/locale/es_ES';
 import frFR from 'antd/lib/locale/fr_FR';
 import { getEnvVarByKey } from 'helpers/EnvVariables';
 import AuthMiddleware from 'middleware/AuthMiddleware';
@@ -25,6 +26,7 @@ import { FENCE_NAMES } from 'common/fenceTypes';
 import ErrorBoundary from 'components/ErrorBoundary';
 import Spinner from 'components/uiKit/Spinner';
 import NotificationContextHolder from 'components/utils/NotificationContextHolder';
+import useFeatureToggle from 'hooks/useFeatureToggle';
 import { initGa } from 'services/analytics';
 import { useLang } from 'store/global';
 import { DYNAMIC_ROUTES, STATIC_ROUTES } from 'utils/routes';
@@ -44,13 +46,29 @@ const VariantEntity = loadable(() => import('views/VariantEntity'), loadableProp
 const FileEntity = loadable(() => import('views/FileEntity'), loadableProps);
 const ParticipantEntity = loadable(() => import('views/ParticipantEntity'), loadableProps);
 const ProfileSettings = loadable(() => import('views/ProfileSettings'), loadableProps);
+const SetOperations = loadable(() => import('views/Analytics/SetOperations'), loadableProps);
+
+const FT_SET_OPERATIONS = 'ANALYTICS_SET_OPERATIONS';
 
 initGa();
+
+const getLocale = (lang: string) => {
+  switch (lang) {
+    case LANG.FR:
+      return frFR;
+    case LANG.ES:
+      return esES;
+    default:
+      return enUS;
+  }
+};
 
 const App = () => {
   const lang = useLang();
   const { keycloak, initialized } = useKeycloak();
   const keycloakIsReady = keycloak && initialized;
+
+  const { isEnabled: isSetOperationsEnabled } = useFeatureToggle(FT_SET_OPERATIONS);
 
   setLocale(lang);
 
@@ -64,10 +82,7 @@ const App = () => {
   }
 
   return (
-    <ConfigProvider
-      locale={lang === LANG.FR ? frFR : enUS}
-      renderEmpty={() => <Empty imageType="grid" />}
-    >
+    <ConfigProvider locale={getLocale(lang)} renderEmpty={() => <Empty imageType="grid" />}>
       <ApolloProvider backend={GraphqlBackend.ARRANGER}>
         <div className="App" id="appContainer">
           {keycloakIsReady ? (
@@ -191,6 +206,17 @@ const App = () => {
                       </ProtectedRoute>
                     }
                   />
+
+                  {isSetOperationsEnabled && (
+                    <Route
+                      path={STATIC_ROUTES.ANALYTICS_SET_OPERATIONS}
+                      element={
+                        <ProtectedRoute>
+                          <SetOperations />
+                        </ProtectedRoute>
+                      }
+                    />
+                  )}
 
                   <Route path="*" element={<Navigate to={STATIC_ROUTES.DASHBOARD} />} />
                 </Routes>