+++ src/services/analytics.ts
@@ -264,3 +264,21 @@ export const trackVennViewEntityCounts = (type: string, entityCount: number) => 
     });
   }
 };
+
+export const trackLaunchSetOperations = () => {
+  if (isGaActive) {
+    ReactGA.event({
+      category: 'Analytics',
+      action: 'Analytics - Set Operations Launch',
+    });
+  }
+};
+
+export const trackCompareSetOperations = () => {
+  if (isGaActive) {
+    ReactGA.event({
+      category: 'Analytics',
+      action: 'Analytics - Set Operations Compare',
+    });
+  }
+};